reverted back its in subject

+++ cpp_projs/CPP06/ex01/Serialization.cpp
@@ -4,17 +4,17 @@ Serialization::Serialization(){};
 
 Serialization::~Serialization(){};
 
-// 1. unsigned long is an unsigned integer type
+// 1. uintptr_t is an unsigned integer type
 /* 2. reinterpret cast performs a low-level, unsafe conversion between unrelated pointer types
 or between a pointer and an integer. tells compiler to reinterpret the raw bit pattern of a variable as
 a different type */
 
-unsigned long Serialization::serialize(Data *ptr)
+uintptr_t Serialization::serialize(Data *ptr)
 {
     return reinterpret_cast<uintptr_t>(ptr);
 };
 
-Data* Serialization::deserialize(unsigned long raw)
+Data* Serialization::deserialize(uintptr_t raw)
 {
     return reinterpret_cast<Data *>(raw);
 };
