+++ cpp_projs/CPP06/ex01/Serialization.hpp
@@ -11,8 +11,8 @@ class Serialization {
         Serialization();
         ~Serialization();
     public:
-        static unsigned long serialize(Data *ptr);
-        static Data* deserialize(unsigned long raw);
+        static uintptr_t serialize(Data *ptr);
+        static Data* deserialize(uintptr_t raw);
 };
 
 #endif
